+++ Parada.h
@@ -12,8 +12,8 @@ class Parada
 {
 	private:
 		int ID, codigo, nv; //nv = número de veículos que passam pela parada
-		float latitude;
-		float longitude;
+		double latitude;
+		double longitude;
 		char terminal;
 		Parada* prox;  // Este atributo deve ser usado somente se for implementada uma lista ENCADEADA de paradas
 	public:
@@ -22,16 +22,16 @@ class Parada
 		int get_ID();
 		int get_codigo();
 		int get_nv();
-		float get_latitude();
-		float get_longitude();
+		double get_latitude();
+		double get_longitude();
 		char get_terminal();
 		Parada* get_prox();
 		void exibir_listaVeiculosQuePassamPorAqui();
 		void set_ID(int _ID);
 		void set_codigo(int _codigo);
 		void set_nv(int _nv);
-		void set_latitude(float _latitude);
-		void set_longitude(float _longitude);
+		void set_latitude(double _latitude);
+		void set_longitude(double _longitude);
 		void set_terminal(char _terminal);
 		void set_prox(Parada *p);
 		void RegistraVeiculoNaParada(Veiculo *v);
